requirements.txt
asgiref==3.5.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
docopt==0.6.2
gunicorn==20.1.0
idna==3.3
Markdown==3.3.6
Pillow==9.1.1
pipreqs==0.4.11
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.1.0
yarg==0.1.9
